requirements.txt
torch==1.5.0
torchvision==0.6.0
tqdm
Shapely==1.7.0
Pillow==7.1.2
opencv-python==4.2.0.34
opencv-contrib-python==4.2.0.34
PyYAML
matplotlib